+++ src/main/java/mori/task/interview/tariffComparison/model/PackagedTarrif.java
@@ -4,12 +4,12 @@ import java.math.BigDecimal;
 
 public class PackagedTarrif extends Product {
 
-    public PackagedTarrif(int consumption) {
+    public PackagedTarrif(double consumption) {
         super("Packaged Tariff", consumption);
     }
 
     @Override
-    public BigDecimal calculationModel(int consumption) {
+    public BigDecimal calculationModel(double consumption) {
         if (consumption <= 4000) {
             return BigDecimal.valueOf(800.0);
         } else {
